+++ resources/js/components/DetailField.vue
@@ -4,7 +4,7 @@
 
             <a 
                 class="inline-block font-bold cursor-pointer mr-2 animate-text-color select-none" 
-                :class="{ 'text-60': localeKey !== currentLocale, 'text-primary border-b-2': localeKey === currentLocale }"
+                :class="{ 'text-60': localeKey !== currentLocale, 'text-primary': localeKey === currentLocale }"
                 :key="`a-${localeKey}`" 
                 v-for="(locale, localeKey) in field.locales"
                 @click="changeTab(localeKey)"